sysalmacenfx

+++ POO2024C2G1-devSDB/CalcFX/src/main/java/pe/edu/upeu/calcfx/servicio/CalcServiceI.java
@@ -1,4 +1,3 @@
-
 package pe.edu.upeu.calcfx.servicio;
 
 import pe.edu.upeu.calcfx.modelo.CalcTO;
@@ -8,8 +7,8 @@ public interface CalcServiceI {
 
     public void guardarResultados(CalcTO to);//C
     public List<CalcTO> obtenerResultados();//R
-    public void actualizarResultados(CalcTO to, int index);//U
-    public void eliminarResultados(int index);//D
+    public void actualizarResultados(CalcTO to, Long index);//U
+    public void eliminarResultados(Long index);//D
 
 
 }
